pyproject.toml
[tool.poetry]
name = "plutusfi-mm-backend"
version = "0.0.1"
description = ""
authors = ["Carlos Eduardo <suportebeloj@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.13.1"
black = "^22.6.0"
Django = "^4.0.6"
django-environ = "^0.9.0"
mysqlclient = "^2.1.1"
PyMySQL = "^1.0.2"
Faker = "^13.15.1"
factory-boy = "^3.2.1"
cryptography = "^37.0.4"
djangorestframework-simplejwt = "^5.2.0"
django-cors-headers = "^3.13.0"
gunicorn = "^20.1.0"
whitenoise = "^6.2.0"
Pillow = "^9.2.0"
pandas = "^1.4.3"
dj-inmemorystorage = "^2.1.0"
requests = "^2.28.1"
drf-yasg = "^1.21.3"

[tool.poetry.dev-dependencies]
ipython = "^8.4.0"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
